created internal package for common test structs and functions
