+++ test/pkg/source/internal/testcommons.go
@@ -1,0 +1,47 @@
+package internal
+
+import (
+	"context"
+	"fmt"
+	"github.com/docker/go-connections/nat"
+	"github.com/testcontainers/testcontainers-go"
+)
+
+var ResticReq = testcontainers.ContainerRequest{
+	Image:        "restic/rest-server:latest",
+	ExposedPorts: []string{"8000/tcp"},
+	Env: map[string]string{
+		"OPTIONS":         "--no-auth",
+		"RESTIC_PASSWORD": "mongorepo",
+	},
+}
+
+type TestContainerSetup struct {
+	Container testcontainers.Container
+	Address   string
+	Port      string
+}
+
+func NewTestContainerSetup(ctx context.Context, request *testcontainers.ContainerRequest, port nat.Port) (TestContainerSetup, error) {
+	result := TestContainerSetup{}
+	container, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
+		ContainerRequest: *request,
+		Started:          true,
+	})
+	if err != nil {
+		return TestContainerSetup{}, err
+	}
+	result.Container = container
+	contPort, err := container.MappedPort(ctx, port)
+	if err != nil {
+		return TestContainerSetup{}, err
+	}
+	result.Port = fmt.Sprint(contPort.Int())
+	host, err := container.Host(ctx)
+	if err != nil {
+		return TestContainerSetup{}, err
+	}
+	result.Address = host
+
+	return result, nil
+}
