+++ test/pkg/source/pgdump/pgdump_test.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"github.com/mittwald/brudi/pkg/source"
+	commons "github.com/mittwald/brudi/test/pkg/source/internal"
 
 	"github.com/docker/go-connections/nat"
 	_ "github.com/jackc/pgx/stdlib"
@@ -44,21 +45,6 @@ var pgRequest = testcontainers.ContainerRequest{
 	WaitingFor: wait.ForLog("database system is ready to accept connections"),
 }
 
-var resticReq = testcontainers.ContainerRequest{
-	Image:        "restic/rest-server:latest",
-	ExposedPorts: []string{"8000/tcp"},
-	Env: map[string]string{
-		"OPTIONS":         "--no-auth",
-		"RESTIC_PASSWORD": "mongorepo",
-	},
-}
-
-type TestContainerSetup struct {
-	Container testcontainers.Container
-	Address   string
-	Port      string
-}
-
 func (pgDumpTestSuite *PGDumpTestSuite) SetupTest() {
 	viper.Reset()
 	viper.SetConfigType("yaml")
@@ -70,32 +56,8 @@ func (pgDumpTestSuite *PGDumpTestSuite) TearDownTest() {
 	viper.Reset()
 }
 
-func newTestContainerSetup(ctx context.Context, request *testcontainers.ContainerRequest, port nat.Port) (TestContainerSetup, error) {
-	result := TestContainerSetup{}
-	container, err := testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
-		ContainerRequest: *request,
-		Started:          true,
-	})
-	if err != nil {
-		return TestContainerSetup{}, err
-	}
-	result.Container = container
-	contPort, err := container.MappedPort(ctx, port)
-	if err != nil {
-		return TestContainerSetup{}, err
-	}
-	result.Port = fmt.Sprint(contPort.Int())
-	host, err := container.Host(ctx)
-	if err != nil {
-		return TestContainerSetup{}, err
-	}
-	result.Address = host
-
-	return result, nil
-}
-
 // createMongoConfig creates a brudi config for the mongodump command
-func createPGConfig(container TestContainerSetup, useRestic bool, resticIP, resticPort string) []byte {
+func createPGConfig(container commons.TestContainerSetup, useRestic bool, resticIP, resticPort string) []byte {
 	if !useRestic {
 		return []byte(fmt.Sprintf(`
 pgdump:
@@ -178,7 +140,7 @@ func (pgDumpTestSuite *PGDumpTestSuite) TestBasicPGDump() {
 	pgDumpTestSuite.Require().NoError(err)
 
 	// create a mysql container to test backup function
-	pgBackupTarget, err := newTestContainerSetup(ctx, &pgRequest, port)
+	pgBackupTarget, err := commons.NewTestContainerSetup(ctx, &pgRequest, port)
 	pgDumpTestSuite.Require().NoError(err)
 	// connect to mysql database using the driver
 	connectionString := fmt.Sprintf("user=postgresuser password=postgresroot host=%s port=%s database=%s sslmode=disable",
@@ -214,7 +176,7 @@ func (pgDumpTestSuite *PGDumpTestSuite) TestBasicPGDump() {
 	pgDumpTestSuite.Require().NoError(err)
 
 	// setup second pgsql container to test if correct data is restored
-	pgRestoreTarget, err := newTestContainerSetup(ctx, &pgRequest, port)
+	pgRestoreTarget, err := commons.NewTestContainerSetup(ctx, &pgRequest, port)
 	pgDumpTestSuite.Require().NoError(err)
 
 	connectionString2 := fmt.Sprintf("user=postgresuser password=postgresroot host=%s port=%s database=%s sslmode=disable",
@@ -253,11 +215,11 @@ func (pgDumpTestSuite *PGDumpTestSuite) TestPGDumpRestic() {
 	pgDumpTestSuite.Require().NoError(err)
 
 	// setup a container running the restic rest-server
-	resticContainer, err := newTestContainerSetup(ctx, &resticReq, "8000/tcp")
+	resticContainer, err := commons.NewTestContainerSetup(ctx, &commons.ResticReq, "8000/tcp")
 	pgDumpTestSuite.Require().NoError(err)
 
 	// create a mysql container to test backup function
-	pgBackupTarget, err := newTestContainerSetup(ctx, &pgRequest, port)
+	pgBackupTarget, err := commons.NewTestContainerSetup(ctx, &pgRequest, port)
 	pgDumpTestSuite.Require().NoError(err)
 	// connect to mysql database using the driver
 	connectionString := fmt.Sprintf("user=postgresuser password=postgresroot host=%s port=%s database=%s sslmode=disable",
@@ -293,7 +255,7 @@ func (pgDumpTestSuite *PGDumpTestSuite) TestPGDumpRestic() {
 	pgDumpTestSuite.Require().NoError(err)
 
 	// setup second pgsql container to test if correct data is restored
-	pgRestoreTarget, err := newTestContainerSetup(ctx, &pgRequest, port)
+	pgRestoreTarget, err := commons.NewTestContainerSetup(ctx, &pgRequest, port)
 	pgDumpTestSuite.Require().NoError(err)
 
 	connectionString2 := fmt.Sprintf("user=postgresuser password=postgresroot host=%s port=%s database=%s sslmode=disable",
